+++ tiberiandawn/infantry.h
@@ -162,7 +162,7 @@ public:
     **	object interacts with the map and thus indirectly controls rendering.
     */
     virtual short const* Overlap_List(void) const;
-    virtual void Get_AABB(int& dx0, int& dy0, int& dx1, int& dy1) const;
+    virtual void Get_AABB(short& dx0, short& dy0, short& dx1, short& dy1) const;
     virtual void Draw_It(int x, int y);
     virtual void Look(bool incremental = false);
