+++ tiberiandawn/terrain.cpp
@@ -352,7 +352,7 @@ bool TerrainClass::Mark(MarkType mark)
  *   06/27/1994 JLB : Created.                                                                 *
  *   11/09/1994 JLB : Changed selected terrain highlight method.                               *
  *=============================================================================================*/
-void TerrainClass::Get_AABB(int& x0, int& y0, int& x1, int& y1) const
+void TerrainClass::Get_AABB(short& x0, short& y0, short& x1, short& y1) const
 {
     int const x = Coord_X(Coord);
     int const y = Coord_Y(Coord);
